broke terraform install logic into another file
